Protecting crud routes

+++ app/controllers/posts_controller.rb
@@ -1,5 +1,7 @@
 class PostsController < ApplicationController
 
+  before_action :logged_in?, except: [:index, :show]
+
   def index
 
   end
@@ -41,21 +43,24 @@ class PostsController < ApplicationController
     @post = Post.find_by(id: post_id)
     city_id = params[:city_id]
     @city = City.find_by(id: city_id)
-    render :edit
+    if session[:user_id] != @post.user_id
+      redirect_to login_path
+    end
   end
 
   def update
     return if inactive_redirect
     city = City.find(params[:city_id])
     post_id = params[:id]
-    post = Post.find_by(id: post_id)
+    @post = Post.find_by(id: post_id)
     user_id = current_user[:id]
-    post[:user_id] = user_id
-    if post.update(post_params)
+    @post[:user_id] = user_id
+    if session[:user_id] == @post.user_id
+      @post.update(post_params)
       flash[:notice] = "Post successfully updated!"
       redirect_to city_path(city)
     else
-      flash[:error] = post.errors.full_messages.join(", ")
+      flash[:error] = @post.errors.full_messages.join(", ")
       redirect_to edit_post_path
     end
   end
@@ -64,10 +69,17 @@ class PostsController < ApplicationController
     return if inactive_redirect
     post_id = params[:id]
     post = Post.find_by(id: post_id)
-    post.destroy
-    city_id = params[:city_id]
-    city = City.find_by(id: city_id)
-    redirect_to city_path(city)
+
+
+    if session[:user_id] == post.user_id
+      post.destroy
+      city_id = params[:city_id]
+      city = City.find_by(id: city_id)
+      redirect_to city_path(city)
+    else
+      flash[:notice] = "You may only delete your own posts."
+      redirect_to login_path
+    end
   end
 
   private
